Added interface to link pair
A storage container type to hold an interface
and the link connected to it. This will be useful
for connecting and disconnecting links to devices
with multiple interfaces (will work for single
interface devices as well). Devices should use
a vector of NetLinkPairs to store all interfaces
and links that pertains to it.