Removed get_addr from connection interface
This interface is made to be used by the link
and the link doesn't need to know the address,
it only need to know to which device to relay
the message (this will of course be the one
that the message did not come from). Some
mode of detecting this needs to be implemented
into the link.